# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/feedback/{id}": {
            "get": {
                "description": "Возвращает информацию об опросе по её идентификатору",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Опрос"
                ],
                "summary": "Получение опроса по идентификатору",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор опроса",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация об опросе",
                        "schema": {
                            "$ref": "#/definitions/model.FeedbackGetResponse"
                        }
                    },
                    "400": {
                        "description": "Недопустимый идентификатор опроса",
                        "schema": {
                            "$ref": "#/definitions/model.FeedbackGetResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.FeedbackGetResponse"
                        }
                    }
                }
            }
        },
        "/feedback/{id}/delete": {
            "delete": {
                "description": "Удаляет опрос по её идентификатору",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Опрос"
                ],
                "summary": "Удаление опроса",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор опроса",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Начало даты формирования",
                        "name": "startFormationDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Конец даты формирования",
                        "name": "endFormationDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Статус опроса",
                        "name": "feedbackStatus",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список групп",
                        "schema": {
                            "$ref": "#/definitions/model.FeedbackRequest"
                        }
                    },
                    "400": {
                        "description": "Недопустимый идентификатор опроса",
                        "schema": {
                            "$ref": "#/definitions/model.FeedbackRequest"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.FeedbackRequest"
                        }
                    }
                }
            }
        },
        "/feedback/{id}/status": {
            "put": {
                "description": "Обновляет статус опроса для модератора по идентификатору опроса",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Опрос"
                ],
                "summary": "Обновление статуса опроса для модератора",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор опроса",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новый статус опроса",
                        "name": "feedbackStatus",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.FeedbackUpdateStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о доставке",
                        "schema": {
                            "$ref": "#/definitions/model.FeedbackGetResponse"
                        }
                    },
                    "400": {
                        "description": "Недопустимый идентификатор опроса или ошибка чтения JSON объекта",
                        "schema": {
                            "$ref": "#/definitions/model.FeedbackGetResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.FeedbackGetResponse"
                        }
                    }
                }
            }
        },
        "/feedback/{id}/user": {
            "put": {
                "description": "Обновляет статус опроса для пользователя по идентификатору опроса",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Опрос"
                ],
                "summary": "Обновление статуса опроса для пользователя",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор опроса",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о доставке",
                        "schema": {
                            "$ref": "#/definitions/model.FeedbackGetResponse"
                        }
                    },
                    "400": {
                        "description": "Недопустимый идентификатор опроса",
                        "schema": {
                            "$ref": "#/definitions/model.FeedbackGetResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.FeedbackGetResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.FeedbackGetResponse": {
            "type": "object",
            "properties": {
                "completion_date": {
                    "type": "string"
                },
                "creation_date": {
                    "type": "string"
                },
                "feedback_id": {
                    "type": "integer"
                },
                "feedback_status": {
                    "type": "string"
                },
                "formation_date": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Group"
                    }
                }
            }
        },
        "model.FeedbackRequest": {
            "type": "object",
            "properties": {
                "completion_date": {
                    "type": "string"
                },
                "creation_date": {
                    "type": "string"
                },
                "feedback_id": {
                    "type": "integer"
                },
                "feedback_status": {
                    "type": "string"
                },
                "formation_date": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                }
            }
        },
        "model.FeedbackUpdateStatusRequest": {
            "type": "object",
            "properties": {
                "feedback_status": {
                    "type": "string"
                }
            }
        },
        "model.Group": {
            "type": "object",
            "properties": {
                "contacts": {
                    "type": "string",
                    "example": "+7(999)999-99-99"
                },
                "course": {
                    "type": "integer",
                    "example": 1
                },
                "group_code": {
                    "type": "string",
                    "example": "RT5-51B"
                },
                "group_id": {
                    "type": "integer"
                },
                "group_status": {
                    "type": "string",
                    "example": "обучается"
                },
                "photo": {
                    "type": "string",
                    "example": "http://example.com/group.jpg"
                },
                "students": {
                    "type": "integer",
                    "example": 23
                }
            }
        },
...
    }
}